imagegen3/requirements.txt
vertexai
Flask
Pillow
google-cloud-aiplatform
